Updated file references moved all of the decorations to decorations.dart
