+++ app/integration_test/account_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:app/main.dart' as app;
 import 'test_utils.dart';
 
 void main() async {
