+++ app/integration_test/swipe_left_to_delete_test.dart
@@ -1,7 +1,6 @@
 import 'package:app/provider/entry.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:app/pages/entries.dart';
 import 'package:app/provider/settings.dart' as settings;
 import 'package:app/main.dart' as app;
